Dispatch math related methods to helpers.math
