+++ src/core/core.ticks.js
@@ -42,13 +42,13 @@ module.exports = {
 				}
 			}
 
-			var logDelta = helpers.log10(Math.abs(delta));
+			var logDelta = helpers.math.log10(Math.abs(delta));
 			var tickString = '';
 
 			if (tickValue !== 0) {
 				var maxTick = Math.max(Math.abs(ticks[0]), Math.abs(ticks[ticks.length - 1]));
 				if (maxTick < 1e-4) { // all ticks are small numbers; use scientific notation
-					var logTick = helpers.log10(Math.abs(tickValue));
+					var logTick = helpers.math.log10(Math.abs(tickValue));
 					tickString = tickValue.toExponential(Math.floor(logTick) - Math.floor(logDelta));
 				} else {
 					var numDecimal = -1 * Math.floor(logDelta);
@@ -63,7 +63,7 @@ module.exports = {
 		},
 
 		logarithmic: function(tickValue, index, ticks) {
-			var remain = tickValue / (Math.pow(10, Math.floor(helpers.log10(tickValue))));
+			var remain = tickValue / (Math.pow(10, Math.floor(helpers.math.log10(tickValue))));
 
 			if (tickValue === 0) {
 				return '0';
