+++ src/core/core.tooltip.js
@@ -146,7 +146,7 @@ var positioners = {
 			var el = elements[i];
 			if (el && el.hasValue()) {
 				var center = el.getCenterPoint();
-				var d = helpers.distanceBetweenPoints(eventPosition, center);
+				var d = helpers.math.distanceBetweenPoints(eventPosition, center);
 
 				if (d < minDistance) {
 					minDistance = d;
