+++ src/helpers/index.js
@@ -4,3 +4,4 @@ module.exports = require('./helpers.core');
 module.exports.easing = require('./helpers.easing');
 module.exports.canvas = require('./helpers.canvas');
 module.exports.options = require('./helpers.options');
+module.exports.math = require('./helpers.math');
