+++ src/scales/scale.radialLinear.js
@@ -157,7 +157,7 @@ function fitWithPointLabels(scale) {
 
 		// Add quarter circle to make degree 0 mean top of circle
 		var angleRadians = scale.getIndexAngle(i);
-		var angle = helpers.toDegrees(angleRadians) % 360;
+		var angle = helpers.math.toDegrees(angleRadians) % 360;
 		var hLimits = determineLimits(angle, pointPosition.x, textSize.w, 0, 180);
 		var vLimits = determineLimits(angle, pointPosition.y, textSize.h, 90, 270);
 
@@ -262,7 +262,7 @@ function drawPointLabels(scale) {
 			ctx.fillStyle = pointLabelFontColor;
 
 			var angleRadians = scale.getIndexAngle(i);
-			var angle = helpers.toDegrees(angleRadians);
+			var angle = helpers.math.toDegrees(angleRadians);
 			ctx.textAlign = getTextAlignForAngle(angle);
 			adjustPointPositionForLabelHeight(angle, scale._pointLabelSizes[i], pointLabelPosition);
 			fillText(ctx, scale.pointLabels[i] || '', pointLabelPosition, plFont.lineHeight);
@@ -311,7 +311,7 @@ function drawRadiusLine(scale, gridLineOpts, radius, index) {
 }
 
 function numberOrZero(param) {
-	return helpers.isNumber(param) ? param : 0;
+	return helpers.math.isNumber(param) ? param : 0;
 }
 
 module.exports = LinearScaleBase.extend({
